:truck: Separate mode 1 render logic
