+++ ppu/src/lib.rs
@@ -3,4 +3,5 @@ pub mod vram;
 pub mod cgram;
 pub mod ppu;
 pub mod registers;
-pub mod renderer;
+
+pub mod rendering;
